# apps/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: FlowForge Control Plane API
  version: 0.13.0
  description: |
    Versioned control-plane API. Conventions:

    - Correlation: every response sets `X-Request-ID`. A caller value is
      accepted only when it is 16–128 ASCII letters, digits, or hyphens;
      otherwise the API generates one. The same id is returned in problem
      documents as `request_id` and in structured request logs.
    - Errors: 4xx/5xx responses use RFC 9457 `application/problem+json`
      (`type`, `title`, `status`, `detail`, `instance`) plus `code` and
      `request_id`. Problem details never echo request bodies, credentials,
      or other secret material.
      - Invalid requests use `invalid-request` (400) or `request-too-large` (413).
      Invalid workflow YAML uses `invalid-workflow` (400) with an `errors`
      array (path, line, column, code, message). Authenticated-but-failed
      flows use `unauthenticated` (401) and `forbidden` (403). Unique
      identity collisions, draft revision mismatches, and duplicate
      published digests use `conflict` (409).
    - Browser clients authenticate with `ff_session` / `ff_csrf` cookies from
      `POST /session`. State-changing cookie requests require `X-CSRF-Token`.
      Non-browser callers may keep using `X-FlowForge-Issuer` /
      `X-FlowForge-Subject`. A session cookie wins; conflicting identity
      headers fail closed.
    - Workspace identity is server-derived from `(tenant_id|tenant_slug,
...
servers:
...
paths:
...
  /scripts/catalog:
    get:
      operationId: getScriptCatalog
      summary: Script engine catalog
      description: |
        Requires `opsconfig.view`. Documents `script.python` / `script.go`
        node fields, publish/scan/sign/pin rules, runtime-profile digest
        pinning, E9.2 isolation guarantees (UID/FS/caps/no_new_privs/egress),
        E9.3 typed I/O + retry/indeterminate contract (io, retry.ui, retry.probe,
        scoped handles, allowlisted env), and E9.4 revocation + emergency-stop
        (`revocation`, `emergencyStop`, error codes). Does not start a runner.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
      responses:
        "200":
          description: Script engine catalog
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptEngineCatalog"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /scripts/{artifactId}:
    get:
      operationId: getScriptArtifact
      summary: Get a published script artifact
      description: |
        Requires `workflow.view`. Returns metadata, digest, scan status, and
        signature. Never includes the package blob or storage locator.
        Cross-workspace IDs are `404`.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
        - name: artifactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Artifact metadata
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptArtifact"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /scripts/{artifactId}/revoke:
    post:
      operationId: revokeScriptArtifact
      summary: Revoke a published script artifact
      description: |
        Requires `script.revoke` (operator/admin). Sets `revokedAt` (idempotent).
        Revoked artifacts cannot start; start/claim/heartbeat-before-dispatch
        and `VerifyForDispatch` fail closed with `artifact-revoked`. Never
        returns the package blob. Host-supplied workspace IDs are `400`.
        Cross-workspace IDs are `404`. Audit `script.artifact.revoke` is
        secret-free.
      parameters:
        - $ref: "#/components/parameters/RequestID"
        - $ref: "#/components/parameters/Issuer"
        - $ref: "#/components/parameters/Subject"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/TenantID"
        - $ref: "#/components/parameters/TenantSlug"
        - $ref: "#/components/parameters/WorkbenchKey"
        - $ref: "#/components/parameters/HostWorkspaceID"
        - name: artifactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptRevokeRequest"
      responses:
        "200":
          description: Revoked artifact metadata
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptArtifact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Correlation identifier. Accepted only when 16–128 ASCII letters, digits, or hyphens.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: ^[A-Za-z0-9-]{16,128}$
    Issuer:
      name: X-FlowForge-Issuer
      in: header
      required: false
      description: |
        Host identity issuer for non-browser callers. Browser clients should
        use `ff_session` instead. Ignored when a session cookie is present
        unless it conflicts (conflict fails closed). Does not authorize a workspace.
      schema:
        type: string
    Subject:
      name: X-FlowForge-Subject
      in: header
      required: false
      description: Host identity subject. Unique with issuer. Does not authorize a workspace.
      schema:
        type: string
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: |
        Required on state-changing requests authenticated by `ff_session`.
        Must match the `ff_csrf` cookie and the server-side session secret.
      schema:
        type: string
    TenantID:
      name: X-FlowForge-Tenant-ID
      in: header
      required: false
      description: Tenant UUID used with workbench_key to resolve workspace identity.
      schema:
        type: string
        format: uuid
    TenantSlug:
      name: X-FlowForge-Tenant-Slug
      in: header
      required: false
      description: Tenant slug used with workbench_key when tenant id is omitted.
      schema:
        type: string
    WorkbenchKey:
      name: X-FlowForge-Workbench-Key
      in: header
      required: false
      description: Host workbench key. Combined with tenant identity, this is the only accepted workspace lookup.
      schema:
        type: string
    HostWorkspaceID:
      name: X-FlowForge-Workspace-ID
      in: header
      required: false
      description: Untrusted host-supplied workspace UUID. Rejected unless it matches the server-derived `(tenant, workbench_key)` workspace.
      schema:
        type: string
        format: uuid
...
  headers:
    XRequestID:
      description: Correlation identifier. Caller values are accepted only when 16–128 ASCII letters, digits, or hyphens.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: ^[A-Za-z0-9-]{16,128}$
  schemas:
...
    ScriptEngineCatalog:
      type: object
      required: [languages, publishRules, nodes, errors, permissions, isolation]
      properties:
        languages:
          type: array
          items: {type: string, enum: [python, go]}
        publishRules:
          type: object
          properties:
            requiredWith:
              type: array
              items: {type: string}
            allowedLanguages:
              type: array
              items: {type: string}
            sourceVisibleInYAML: {type: boolean}
            secretsForbiddenInYAML: {type: boolean}
            draftsCannotExecute: {type: boolean}
            mutableArtifactsRejected: {type: boolean}
            unscannedRejected: {type: boolean}
            unsignedRejected: {type: boolean}
            failedScanRejected: {type: boolean}
            publishedRevisionsPinned: {type: boolean}
            digestPinnedRuntime: {type: boolean}
            maxSourceBytes: {type: integer}
            maxTimeoutSeconds: {type: integer}
        nodes:
          type: array
          items:
            type: object
            properties:
              type: {type: string}
              language: {type: string}
              title: {type: string}
              description: {type: string}
              permissions:
                type: array
                items: {type: string}
              requiredWith:
                type: array
                items: {type: string}
              allowedWith:
                type: array
                items:
                  type: object
                  properties:
                    name: {type: string}
                    kind: {type: string}
                    required: {type: boolean}
                    description: {type: string}
              outputs:
                type: array
                items: {type: string}
              sideEffects: {type: boolean}
              retrySafe: {type: boolean}
              idempotent: {type: boolean}
        errors:
          type: array
          items:
            type: object
            properties:
              code: {type: string}
              status: {type: integer}
              meaning: {type: string}
        permissions:
          type: array
          items: {type: string}
        isolation:
          type: object
          properties:
            nonRoot: {type: boolean}
            readOnlyRootFS: {type: boolean}
            droppedCapabilities: {type: boolean}
            noNewPrivs: {type: boolean}
            noMetadataService: {type: boolean}
            noHostDockerSocket: {type: boolean}
            runtimePackageInstall: {type: boolean}
            approvedImagesOnly: {type: boolean}
            uid: {type: integer}
            gid: {type: integer}
            ephemeralWorkspace: {type: string}
            dropCapabilityNames:
              type: array
              items: {type: string}
            allowPrivilegeEscalation: {type: boolean}
            noServiceAccountMount: {type: boolean}
            defaultDenyEgress: {type: boolean}
            dnsConstrained: {type: boolean}
            metadataCIDRs:
              type: array
              items: {type: string}
            runtimeProfile:
              type: object
              properties:
                language: {type: string}
                imageDigest: {type: string}
                dependencyLockDigest: {type: string}
                limits:
                  type: array
                  items: {type: string}
                egress: {type: string}
                mutableTagsRejected: {type: boolean}
            ciHarness: {type: string}
            kubernetesManifests:
              type: array
              items: {type: string}
            note: {type: string}
            hooks:
              type: array
              items: {type: string}
        hooks:
          type: object
          additionalProperties: {type: string}
        revocation:
          type: object
          properties:
            permission: {type: string}
            route: {type: string}
            idempotent: {type: boolean}
            blocksNewRuns: {type: boolean}
            recheckOn:
              type: array
              items: {type: string}
            failClosed: {type: boolean}
            errorCode: {type: string}
            auditAction: {type: string}
            auditSecretFree: {type: boolean}
            note: {type: string}
        emergencyStop:
          type: object
          properties:
            permission: {type: string}
            route: {type: string}
            policyGated: {type: boolean}
            missingPolicyAllows: {type: boolean}
            policyAllowField: {type: string}
            uncertainOutcome: {type: string}
            beforeDispatch: {type: string}
            neverAssumeAbsent: {type: boolean}
            states:
              type: array
              items: {type: string}
            auditAction: {type: string}
            auditSecretFree: {type: boolean}
            denyWithoutPermission: {type: string}
            note: {type: string}
    ScriptRevokeRequest:
      type: object
      properties:
        reason: {type: string, maxLength: 256}
...
    ScriptArtifact:
      type: object
      required: [id, language, entrypoint, digest, signature, scanStatus, status, createdAt]
      properties:
        id: {type: string, format: uuid}
        language: {type: string, enum: [python, go]}
        entrypoint: {type: string}
        digest: {type: string}
        signature: {type: string}
        scanStatus: {type: string, enum: [pending, clean, failed, unsigned]}
        status: {type: string, enum: [draft, published]}
        runtimeProfileId: {type: string, format: uuid}
        runtimeProfileVersionId: {type: string, format: uuid}
        runtimeProfileDigest: {type: string}
        sourceBytes: {type: integer}
        metadata: {type: object, additionalProperties: true}
        createdBy: {type: string}
        createdAt: {type: string, format: date-time}
        revokedAt: {type: string, format: date-time}
        revokedBy: {type: string}
...
    Problem:
      type: object
      required: [type, title, status, detail, instance, code, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          enum:
            - invalid-request
            - unauthenticated
            - forbidden
            - not-found
            - conflict
            - method-not-allowed
            - request-too-large
            - internal-error
            - dependency-unavailable
            - invalid-workflow
            - retry-denied
            - artifact-mutable
            - artifact-unscanned
            - artifact-unsigned
            - artifact-scan-failed
        request_id:
          type: string
        errors:
          type: array
          description: Field-level workflow validation errors (invalid-workflow only).
          items:
            $ref: "#/components/schemas/WorkflowFieldError"
...
    WorkflowFieldError:
      type: object
      required: [path, code, message]
      properties:
        path:
          type: string
          example: spec.nodes[0].with.namespace
        line:
          type: integer
        column:
          type: integer
        code:
          type: string
          example: invalid-uuid
        message:
          type: string
...
  responses:
...
    Unauthenticated:
      description: Missing or invalid credentials
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: urn:flowforge:problem:unauthenticated
            title: Unauthenticated
            status: 401
            detail: Authentication is required.
            instance: /api/v1/example
            code: unauthenticated
            request_id: 0123456789abcdef0123456789abcdef
    Forbidden:
      description: Authenticated caller is not authorized
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: urn:flowforge:problem:forbidden
            title: Forbidden
            status: 403
            detail: You are not authorized to perform this action.
            instance: /api/v1/example
            code: forbidden
            request_id: 0123456789abcdef0123456789abcdef
...
    NotFound:
      description: The requested path does not exist
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: urn:flowforge:problem:not-found
            title: Not Found
            status: 404
            detail: The requested path does not exist.
            instance: /api/v1/missing
            code: not-found
            request_id: 0123456789abcdef0123456789abcdef
...